pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = [
    "neuralgcm",
    "neuralgcm.experimental",
    "neuralgcm.experimental.atmosphere",
    "neuralgcm.experimental.core",
    "neuralgcm.experimental.xreader",
    "neuralgcm.legacy",
]
package-data = {neuralgcm = ["data/*"]}

[project]
name = "neuralgcm"
description = "Neural General Circulation Models for Weather and Climate"
version = "1.2.1"  # keep in sync with __version__ in neuralgcm/__init__.py
authors = [
    {name = "Google LLC", email = "noreply@google.com"},
]
readme = "README.md"
license = "Apache-2.0"
requires-python = ">=3.10"
dependencies = [
    "dinosaur>=1.3.3",
    "dm-haiku",
    "gin-config",
    "jax",
    "jaxlib",
    "numpy",
    "optax",
    "pandas",
    "tensorflow_probability",
    "xarray",
]

[project.urls]
source = "https://github.com/neuralgcm/neuralgcm"
documentation = "https://neuralgcm.readthedocs.io"

[project.optional-dependencies]
tests = [
    "absl-py",
    "chex",
    "pytest",
]
experimental = [
    "coordax",
    "einops",
    "etils",
    "fiddle",
    "flax",
    "grain",
    "jax-datetime",
    "jax-solar",
    "orbax",
    "treescope",
]